Остается написать код для сохранения фотографии в папку.
